Introduce linting
This project will be using `golang` as the glue language to write
together the plugin as it is a language the author is interesting
practicing.

To ensure the various go "best practices" are not violated this commit
introduces a linting tool `golangci-lint` and switches on linting to the
highest level.

This is then added as a hook to prevent committing incorrectly formatted
code.